moved all shared variables and functions to the base enemy class
